Refactoring and start button is start stop now
