+++ SkinMK2.py
@@ -425,12 +425,12 @@ class Colors:
             On = Rgb.RED
             Off = Rgb.RED_THIRD
 
-        class Play:
-            On = Rgb.RED
-            Off = Rgb.RED_THIRD
+        class Lock:
+            On = Rgb.PURPLE
+            Off = Rgb.PURPLE_THIRD
 
         class Stop:
-            On = Rgb.RED
+            On = Rgb.GREEN_HALF_PULSE
             Off = Rgb.RED_THIRD
 
         class Mute:
